feat: renamed api file names
M  jest.setup.js
M  src/app/(common)/_notifications/_hooks/NotificationsHook.jsx
M  src/app/candidate/_components/CandidateApplyConfirmation.jsx
M  src/app/candidate/_components/CurriculumForm.jsx
M  src/app/candidate/_hooks/ApplyHook.jsx
M  src/app/candidate/_hooks/AutomatedInfoProcessingMngmentSystemsHook.jsx
M  src/app/candidate/_hooks/ContestsHook.jsx
M  src/app/candidate/_hooks/HomeHook.jsx
M  src/app/committee/(chairman-member)/_constants/CandidateConstants.jsx
M  src/app/committee/(chairman-member)/_hooks/CommissionsHook.jsx
M  src/app/committee/(chairman-member)/_hooks/ExamsHook.jsx
M  src/app/committee/(chairman-member)/_hooks/HomeHook.jsx
M  src/app/doctoral-center/(manager-expert)/_components/CandidatesGrid.jsx
M  src/app/doctoral-center/(manager-expert)/_constants/CandidatesConstants.jsx
M  src/app/doctoral-center/(manager-expert)/_hooks/CandidatesHook.jsx
M  src/app/doctoral-center/(manager-expert)/_hooks/ExamsHook.jsx
M  src/app/doctoral-center/(manager-expert)/_hooks/HomeHook.jsx
M  src/app/doctoral-center/(manager-expert)/_hooks/UnauthorizedUsersHook.jsx
M  src/app/doctoral-center/admin/_hooks/EventManagementHook.jsx
M  src/app/doctoral-center/admin/_hooks/HomeHook.jsx
M  src/app/doctoral-center/admin/_hooks/UnauthorizedUsersHook.jsx
M  src/app/doctoral-center/admin/_hooks/UserManagementHook.jsx
M  src/components/main-layout/common/Header.jsx
M  src/hooks/AuthHook.jsx
R  src/lib/api/candidate.jsx -> src/lib/api/CandidateAPI.jsx
R  src/lib/api/committee.jsx -> src/lib/api/CommitteeAPI.jsx
R  src/lib/api/doctoralCenter.jsx -> src/lib/api/DoctoralCenterAPI.jsx
R  src/lib/api/doctoralCenterAdmin.jsx -> src/lib/api/DoctoralCenterAdminAPI.jsx
R  src/lib/api/file.jsx -> src/lib/api/FileAPI.jsx
R  src/lib/api/logs.jsx -> src/lib/api/LogsAPI.jsx
R  src/lib/api/notification.jsx -> src/lib/api/NotificationAPI.jsx
R  src/lib/api/unauthorized.jsx -> src/lib/api/UnauthorizedAPI.jsx
M  src/lib/helpers/APIWrapper.jsx
