+++ src/app/committee/(chairman-member)/_constants/CandidateConstants.jsx
@@ -1,5 +1,5 @@
 import { renderAvatar } from "@/components/cells-renderers/avatar";
-import FileAPI from "@/lib/api/file";
+import FileAPI from "@/api/FileAPI";
 import Translate from "@/lib/helpers/Translate";
 import { createDataUrl } from "@/lib/helpers/utils";
