+++ src/app/doctoral-center/(manager-expert)/_hooks/UnauthorizedUsersHook.jsx
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useMemo, useState } from "react";
 import UnauthorizedUsers from "@/models/UnauthorizedUsers";
 import APIWrapper from "@/lib/helpers/APIWrapper";
-import DoctoralCenterAPI from "@/lib/api/doctoralCenter";
+import DoctoralCenterAPI from "@/api/DoctoralCenterAPI";
 import { runPeriodically } from "@/lib/helpers/utils";
 import Translate from "@/lib/helpers/Translate";
